Refactored rendering systems into one.
