+++ Humble2/src/Humble2/Asset/AssetImporter.cpp
@@ -6,9 +6,7 @@
 #include "Systems\TransformSystem.h"
 #include "Systems\LinkSystem.h"
 #include "Systems\CameraSystem.h"
-#include "Systems\StaticMeshRenderingSystem.h"
-#include "Systems\SpriteRenderingSystem.h"
-#include "Systems\CompositeRenderingSystem.h"
+#include "Systems\RenderingSystem.h"
 #include "Systems\SoundSystem.h"
 
 namespace HBL2
@@ -451,9 +449,7 @@ namespace HBL2
 		scene->RegisterSystem(new TransformSystem);
 		scene->RegisterSystem(new LinkSystem);
 		scene->RegisterSystem(new CameraSystem, SystemType::Runtime);
-		scene->RegisterSystem(new StaticMeshRenderingSystem);
-		scene->RegisterSystem(new SpriteRenderingSystem);
-		scene->RegisterSystem(new CompositeRenderingSystem);
+		scene->RegisterSystem(new RenderingSystem);
 		scene->RegisterSystem(new SoundSystem, SystemType::Runtime);
 
 		return sceneHandle;
@@ -670,9 +666,7 @@ namespace HBL2
 			scene->RegisterSystem(new TransformSystem);
 			scene->RegisterSystem(new LinkSystem);
 			scene->RegisterSystem(new CameraSystem, SystemType::Runtime);
-			scene->RegisterSystem(new StaticMeshRenderingSystem);
-			scene->RegisterSystem(new SpriteRenderingSystem);
-			scene->RegisterSystem(new CompositeRenderingSystem);
+			scene->RegisterSystem(new RenderingSystem);
 			scene->RegisterSystem(new SoundSystem, SystemType::Runtime);
 
 			asset->Indentifier = sceneHandle.Pack();
