--- Humble2/src/Humble2/Systems/LightSystem.cpp
@@ -1,16 +1,0 @@
-#include "LightSystem.h"
-
-namespace HBL2
-{
-	void LightSystem::OnCreate()
-	{
-	}
-
-	void LightSystem::OnUpdate(float ts)
-	{
-	}
-
-	void LightSystem::OnDestroy()
-	{
-	}
-}
